feat:HNG-2145-settings-page-plus-major-refactor
